Player EF updates

+++ Sadie.Database/Models/Players/PlayerRespect.cs
@@ -2,6 +2,6 @@ namespace Sadie.Database.Models.Players;
 
 public class PlayerRespect
 {
-    public long OriginPlayerId { get; set; }
-    public long TargetPlayerId { get; set; }
+    public int OriginPlayerId { get; set; }
+    public int TargetPlayerId { get; set; }
 }
